Fix in FastCGI package: /var/run is a tmpfs in Ubuntu 12.04 and Fedora
The default location of the hhvm-fastcgi PID file is: /var/run/hhvm/hhvm-fastcgi.pid
The hhvm FastCGI daemon is normaly started as a non privileged user like 'www-data.
Due to Linux file permissions the hhvm daemon needs to have write access to
the basedir /var/run/hhvm.
In recent version of Ubuntu and Fedora, the /var/run folder is mounted as 'tmpfs'
device. /var/run folder is created from scratch on every boot. This leads to the
problem that the non privileged user is not able to create the PID file under
/var/run/hhvm/ which was not created during to boot up process.

To fix this issue create the folder on daemon start up.