Avoid downloading the whole pool of packages each build
Summary:
We download the entire segmemnt of our packages that we use,
every time we build something. Not only is this a waste of bandwidth,
but it's actually getting to be a significant part of our build time.
Most of the data doesn't change day-to-day; instead, have a
semi-persistent copy in `/var/tmp` that we can just rsync diffs to. (The
rsync commands we were using were already basically correct to get this
to work anyways!)

This also disables compression in the sync while we're at it, since
we're mostly copying debs, which are all already compressed; the extra
gzip is just burning CPU.

Test Plan:
Manually ran the rsync commands over a couple of days. See
the repo get pulled up-to-date each time. Make a file in
`/var/tmp/staging`, run the second rsync command with `-n`, see that a
new file would be pushed. In all the testing, actually copy-paste the
commands, to ensure that the variable interpolation, trailing slashes,
etc all work out correctly.

Reviewers: ptarjan, sgolemon, fredemmott

Reviewed By: fredemmott

Differential Revision: https://phabricator.fb.com/D1681930

Signature: t1:1681930:1415990733:95935fd42422a66ae2902cd1d7f4c59b7140c0a5