[packaging] Automatically manage adding LTS releases
Summary:
We have separate releases for the LTS series. The packaging
script should detect and LTS stable release, and add it to both the
current stable series as well as the LTS series. (The former is a no-op
if there is a newer stable series.)

Test Plan:
Hacked script to set a tag for the upcoming `3.3.3` and also
to not push anything to `dl.hhvm.com`. Run script for `3.3.3`, see that
it's added to the LTS release as well.

Reviewers: sgolemon, fredemmott, ptarjan

Reviewed By: ptarjan

Differential Revision: https://phabricator.fb.com/D1769682

Signature: t1:1769682:1420757620:4fc1bbd941cc49593be28133da21eafb1faf7977