Remove arch and rpm files
These have been unmaintained for over a year. If someone wants to pick
up maintainership, I'm more than happy to take a PR for updated build
scripts for these (or any other) distros, but it's counterproductive to
have this stuff hanging around right now. It's still in the history of
course if anyone wants to use it as a base for a new version.